pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "EduTrend AI Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "google-adk>=1.27.2",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.42.0",
]
